test api permission success

+++ app/views.py
@@ -7,7 +7,6 @@ from django.shortcuts import get_object_or_404
 
 from app import models, serializers, permissions
 
-
 class SignUpView(APIView):
 
 	def post(self, request):
@@ -39,39 +38,50 @@ class ProjectView(APIView):
 
 class ProjectViewDetail(APIView):
 
-	permission_classes = [permissions.ContributorOnly]
+	permission_classes = [permissions.ApiPermissions]
 
 	def get(self, request, project_id):
-		project = models.Project.objects.get(pk=project_id)
+		project = get_object_or_404(models.Project, pk=project_id)
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		contributor = get_object_or_404(contributors, user=request.user)
+		self.check_object_permissions(self.request, contributor)
 		serializer = serializers.ProjectDetailSerializer(project)
 		return Response(serializer.data)
 
 	def put(self, request, project_id):
-		project = models.Project.objects.get(pk=project_id)
+		project = get_object_or_404(models.Project, pk=project_id)
+		self.check_object_permissions(self.request, project)
 		serializer = serializers.ProjectViewPostSerializer(project, data=request.data)
 		if serializer.is_valid():
 			serializer.save()
 			return Response(serializer.data)
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-	def delete(self, request, pk):
-		project = models.Project.objects.get(pk=pk)
-		content = {"respose": "project {project.title} deleted"}
+	def delete(self, request, project_id):
+		project = get_object_or_404(models.Project, pk=project_id)
+		self.check_object_permissions(self.request, project)
+		content = {"respose": f"project {project.title} deleted"}
 		project.delete()
 		return Response(content)
 
 
 class ContributorView(APIView):
 
-	permission_classes = [IsAuthenticated]
+	permission_classes = [permissions.ApiPermissions]
 
 	def get(self, request, project_id):
+		project = get_object_or_404(models.Project, pk=project_id)
 		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		self.check_object_permissions(self.request, contributors)
 		serializer = serializers.ContributorGetSerializer(contributors, many=True)
 		return Response(serializer.data)
 	
 	def post(self, request, project_id):
-		project = models.Project.objects.get(pk=project_id)
+		project = get_object_or_404(models.Project, pk=project_id)
+		self.check_object_permissions(self.request, project)
 		serializer = serializers.ContributorAddSerializer(data=request.data)
 		if serializer.is_valid():
 			serializer.save(project_id=project.id)
@@ -79,24 +89,36 @@ class ContributorView(APIView):
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 		
 	def delete(self, request, project_id):
-		project = models.Project.objects.get(pk=project_id)
+		project = get_object_or_404(models.Project, pk=project_id)
+		self.check_object_permissions(self.request, project)
 		contributor = models.Contributor.objects.filter(project_id=project_id).get(
-				username=request.data['username'])
-		contributor_deleted.delete()
+				username=request.data['user'])
+		contributor.delete()
 		content = {"response": f"contributor {request.data['username']} deleted"}
 		return Response(content)
 
 
 class IssueView(APIView):
 
-	permission_classes = [IsAuthenticated]
+	permission_classes = [permissions.ApiPermissions]
 
 	def get(self, request, project_id):
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		self.check_object_permissions(self.request, contributors)
 		issues = models.Issue.objects.all()
 		serializer = serializers.IssueSerializer(issues, many=True)
 		return Response(serializer.data)
 
 	def post(self, request, project_id):
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		else:
+			self.check_object_permissions(self.request, contributors)
 		serializer = serializers.IssuePostSerializer(data=request.data)
 		if serializer.is_valid():
 			serializer.save(project_id=project_id, author=request.user)
@@ -106,9 +128,11 @@ class IssueView(APIView):
 
 class IssueDetail(APIView):
 
-	permission_classes = [IsAuthenticated]
+	permission_classes = [permissions.ApiPermissions]
 
 	def put(self, request, project_id, issue_id):
+		issue = get_object_or_404(get_object_or_404(models.Issue, project_id=project_id), pk=issue_id)
+		self.check_object_permissions(self.request, issue)
 		serializer = serializers.IssuePostSerializer(issue, data=request.data)
 		if serializer.is_valid():
 			serializer.save()
@@ -116,7 +140,8 @@ class IssueDetail(APIView):
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 	def delete(self, request, project_id, issue_id):
-		issue = models.Issue.objects.filter(project_id=project_id).get(pk=issue_id)
+		issue = get_object_or_404(get_object_or_404(models.Issue, project_id=project_id), pk=issue_id)
+		self.check_object_permissions(self.request, issue)
 		issue.delete()
 		content = {"response":"issue deleted"}
 		return Response(content)
@@ -124,15 +149,27 @@ class IssueDetail(APIView):
 
 class Comment(APIView):
 
-	permission_classes = [IsAuthenticated]
+	permission_classes = [permissions.ApiPermissions]
 
 	def get(self, request, project_id, issue_id):
-		comments = models.Comment.objects.filter(issue=issue_id)
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		else:
+			self.check_object_permissions(self.request, contributors)
+		comments = get_object_or_404(models.Comment, issue=issue_id)
 		serializer = serializers.CommentSerializer(comments, many=True)
 		return Response(serializer.data)
 
 	def post(self, request, project_id, issue_id):
-		issue = models.Issue.objects.get(pk=issue_id)
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		else:
+			self.check_object_permissions(self.request, contributors)
+		issue = get_object_or_404(get_object_or_404(models.Issue, project_id=project_id), pk=issue_id)
 		serializer = serializers.CommentPostSerializer(data=request.data)
 		if serializer.is_valid():
 			serializer.save(issue=issue, author=request.user)
@@ -141,15 +178,22 @@ class Comment(APIView):
 
 class CommentDetail(APIView):
 
-	permission_classes = [IsAuthenticated]
+	permission_classes = [permissions.ApiPermissions]
 
 	def get(self, request, project_id, issue_id, comment_id):
-		comment = models.Comment.objects.get(pk=comment_id)
+		contributors = models.Contributor.objects.filter(project_id=project_id)
+		if contributors.count() == 0:
+			content = {"detail": "You do not have permission to perform this action."}
+			return Response(content)
+		else:
+			self.check_object_permissions(self.request, contributors)
+		comment =  get_object_or_404(models.Comment, pk=comment_id)
 		serializer = serializers.CommentSerializer(comment)
 		return Response(serializer.data)
 
 	def put(self, request, project_id, issue_id, comment_id):
-		comment = models.Comment.objects.get(pk=comment_id)
+		comment = get_object_or_404(models.Comment, pk=comment_id)
+		self.check_object_permissions(self.request, comment)
 		serializer = serializers.CommentPostSerializer(comment, data=request.data)
 		if serializer.is_valid():
 			serializer.save(issue=issue, author=request.user)
@@ -157,7 +201,8 @@ class CommentDetail(APIView):
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 	def delete(self, request, project_id, issue_id, comment_id):
-		comment = models.Comment.objects.get(pk=comment_id)
+		comment = get_object_or_404(models.Comment, pk=comment_id)
+		self.check_object_permissions(self.request, comment)
 		comment.delete()
 		content = {"response":"comment deleted"}
 		return Response(content)
